requirements.txt
requests==2.24.0
selectorlib0.16.0
pandas==1.1.3